feat: login 확인

+++ src/main/java/com/example/likelionspringboot/domain/article/article/controller/ArticleController.java
@@ -32,6 +32,10 @@ public class ArticleController {
 
     @GetMapping("/article/write")
     String showWrite() {
+        if (!rq.isLogined()){
+            throw new RuntimeException("로그인 후 이용해주세요.");
+        }
+
         return "article/article/write";
     }
 
@@ -40,6 +44,10 @@ public class ArticleController {
             @NotBlank String title,
             @NotBlank String body
     ) {
+        if (!rq.isLogined()){
+            throw new RuntimeException("로그인 후 이용해주세요.");
+        }
+
         Article article = articleService.write(title, body);
 
         return rq.redirect("/article/list", "%d번 게시물 생성되었습니다.".formatted(article.getId()));
@@ -56,6 +64,10 @@ public class ArticleController {
 
     @GetMapping("/article/delete/{id}")
     String delete(@PathVariable long id) {
+        if (!rq.isLogined()){
+            throw new RuntimeException("로그인 후 이용해주세요.");
+        }
+
         articleService.delete(id);
 
         return rq.redirect("/article/list", "%d번 게시물 삭제되었습니다.".formatted(id));
@@ -63,6 +75,10 @@ public class ArticleController {
 
     @GetMapping("/article/modify/{id}")
     String showModify(Model model, @PathVariable long id) {
+        if (!rq.isLogined()){
+            throw new RuntimeException("로그인 후 이용해주세요.");
+        }
+
         Article article = articleService.findById(id).get();
 
         model.addAttribute("article", article);
@@ -72,6 +88,10 @@ public class ArticleController {
 
     @PostMapping("/article/modify/{id}")
     String modify(@PathVariable long id, @NotBlank String title, @NotBlank String body) {
+        if (!rq.isLogined()){
+            throw new RuntimeException("로그인 후 이용해주세요.");
+        }
+
         articleService.modify(id, title, body);
 
         return rq.redirect("/article/list", "%d번 게시물 수정되었습니다.".formatted(id));
